Create separate files for each dex operation
Nice cleanup for monolithic core.go. Also makes future refactors a bit simpler
