+++ x/dex/keeper/multihop_swap.go
@@ -1,6 +1,8 @@
 package keeper
 
 import (
+	"context"
+	"errors"
 	"fmt"
 
 	sdkerrors "cosmossdk.io/errors"
@@ -17,11 +19,115 @@ type MultihopStep struct {
 	tradePairID        *types.TradePairID
 }
 
-type routeOutput struct {
+type MultiHopRouteOutput struct {
 	write   func()
 	coinOut sdk.Coin
 	route   []string
 	dust    sdk.Coins
+}
+
+// MultiHopSwapCore handles logic for MsgMultihopSwap including bank operations and event emissions.
+func (k Keeper) MultiHopSwapCore(
+	goCtx context.Context,
+	amountIn math.Int,
+	routes []*types.MultiHopRoute,
+	exitLimitPrice math_utils.PrecDec,
+	pickBestRoute bool,
+	callerAddr sdk.AccAddress,
+	receiverAddr sdk.AccAddress,
+) (coinOut sdk.Coin, err error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	bestRoute, initialInCoin, err := k.ExecuteMultiHopSwap(ctx, amountIn, routes, exitLimitPrice, pickBestRoute)
+	if err != nil {
+		return sdk.Coin{}, err
+	}
+
+	bestRoute.write()
+	err = k.bankKeeper.SendCoinsFromAccountToModule(
+		ctx,
+		callerAddr,
+		types.ModuleName,
+		sdk.Coins{initialInCoin},
+	)
+	if err != nil {
+		return sdk.Coin{}, err
+	}
+
+	// send both dust and coinOut to receiver
+	// note that dust can be multiple coins collected from multiple hops.
+	err = k.bankKeeper.SendCoinsFromModuleToAccount(
+		ctx,
+		types.ModuleName,
+		receiverAddr,
+		bestRoute.dust.Add(bestRoute.coinOut),
+	)
+	if err != nil {
+		return sdk.Coin{}, fmt.Errorf("failed to send out coin and dust to the receiver: %w", err)
+	}
+
+	ctx.EventManager().EmitEvent(types.CreateMultihopSwapEvent(
+		callerAddr,
+		receiverAddr,
+		initialInCoin.Denom,
+		bestRoute.coinOut.Denom,
+		initialInCoin.Amount,
+		bestRoute.coinOut.Amount,
+		bestRoute.route,
+		bestRoute.dust,
+	))
+
+	return bestRoute.coinOut, nil
+}
+
+// ExecuteMultiHopSwap handles the core logic for MultiHopSwap -- simulating swap operations across all routes (when applicable)
+// and picking the best route to execute. IT DOES NOT PERFORM ANY BANKING OPERATIONS.
+func (k Keeper) ExecuteMultiHopSwap(
+	ctx sdk.Context,
+	amountIn math.Int,
+	routes []*types.MultiHopRoute,
+	exitLimitPrice math_utils.PrecDec,
+	pickBestRoute bool,
+) (bestRoute MultiHopRouteOutput, initialInCoin sdk.Coin, err error) {
+	var routeErrors []error
+	initialInCoin = sdk.NewCoin(routes[0].Hops[0], amountIn)
+	stepCache := make(map[multihopCacheKey]StepResult)
+
+	bestRoute.coinOut = sdk.Coin{Amount: math.ZeroInt()}
+
+	for _, route := range routes {
+		routeDust, routeCoinOut, writeRoute, err := k.RunMultihopRoute(
+			ctx,
+			*route,
+			initialInCoin,
+			exitLimitPrice,
+			stepCache,
+		)
+		if err != nil {
+			routeErrors = append(routeErrors, err)
+			continue
+		}
+
+		if !pickBestRoute || bestRoute.coinOut.Amount.LT(routeCoinOut.Amount) {
+			bestRoute.coinOut = routeCoinOut
+			bestRoute.write = writeRoute
+			bestRoute.route = route.Hops
+			bestRoute.dust = routeDust
+		}
+		if !pickBestRoute {
+			break
+		}
+	}
+
+	if len(routeErrors) == len(routes) {
+		// All routes have failed
+
+		allErr := errors.Join(append([]error{types.ErrAllMultiHopRoutesFailed}, routeErrors...)...)
+
+		return MultiHopRouteOutput{}, sdk.Coin{}, allErr
+	}
+
+	return bestRoute, initialInCoin, nil
 }
 
 func (k Keeper) HopsToRouteData(
